Fix package tangle - move new template to its own package.
