+++ spring-kafka/src/main/java/org/springframework/kafka/sendreceive/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides classes for request/reply semantics.
+ */
+package org.springframework.kafka.sendreceive;
