+++ spring-kafka/src/test/java/org/springframework/kafka/ReplyingKafkaTemplateTests.java
@@ -48,6 +48,7 @@ import org.springframework.kafka.core.DefaultKafkaProducerFactory;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.kafka.listener.KafkaMessageListenerContainer;
 import org.springframework.kafka.listener.config.ContainerProperties;
+import org.springframework.kafka.sendreceive.ReplyingKafkaTemplate;
 import org.springframework.kafka.support.KafkaHeaders;
 import org.springframework.kafka.support.RequestReplyFuture;
 import org.springframework.kafka.support.SimpleKafkaHeaderMapper;
